requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
sqlalchemy==2.0.23
pydantic==2.4.2
python-dotenv==1.0.0
httpx==0.25.1
python-multipart==0.0.6
aiofiles==23.2.1 